% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADM_SERV.R
\name{adm_serv}
\alias{adm_serv}
\title{Medical Services}
\usage{
adm_serv(col)
}
\arguments{
\item{col}{A column that typically ends in "_SV" or classified using logic ADM_SERV.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Medical Services
}
