% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02 - Age Cleanup.R
\name{age_cleanup}
\alias{age_cleanup}
\title{Calculating Age if AGE is missing using DOB}
\usage{
age_cleanup(age, dob, adm_dt)
}
\arguments{
\item{age}{Raw AGE of Patient}

\item{dob}{Date of Birth of Patient}

\item{adm_dt}{Date of Admission (eda_date_a)}
}
\value{
Returns the AGE of Patient in Years
}
\description{
Calculating Age if AGE is missing using DOB
}
