% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BURN_TYPE.R
\name{burn_type}
\alias{burn_type}
\title{Burn Type}
\usage{
burn_type(col)
}
\arguments{
\item{col}{A column in PTOS data that falls under BURN_TYPE}
}
\value{
It translations of the code into human friendly values.
}
\description{
Burn Type
}
