% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA_DISCGTO.R
\name{ca_discgto}
\alias{ca_discgto}
\title{Child Discharged to
Was the child discharged to a different caregiver than when they were admitted?  (pg. 77)}
\usage{
ca_discgto(col)
}
\arguments{
\item{col}{A column in PTOS data typically CA_DISCGTO}
}
\value{
It translations of the code into human friendly values.
}
\description{
Child Discharged to
Was the child discharged to a different caregiver than when they were admitted?  (pg. 77)
}
