% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMP_LC.R
\name{comp_lc}
\alias{comp_lc}
\title{Diagnosis Location of Complication}
\usage{
comp_lc(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with COMP_ and ends with _LC}
}
\value{
It translations of the code into human friendly values.
}
\description{
Diagnosis Location of Complication
}
