% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLIC.R
\name{complic}
\alias{complic}
\title{Occurances and Complications (pg. 122-125)}
\usage{
complic(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with compr_.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Occurances and Complications (pg. 122-125)
}
