% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNTY.R
\name{county}
\alias{county}
\title{The county of Pennsylvania in which the injury event took place}
\usage{
county(col)
}
\arguments{
\item{col}{Indicate the county or state in which the injury took place using the 2 digit codes  provided.}
}
\value{
State or County translation
}
\description{
The county of Pennsylvania in which the injury event took place
}
