% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Arrival and Discharge.R
\name{create_arrival_discharge_df}
\alias{create_arrival_discharge_df}
\title{Create Normalized Arrival and Discharge Data}
\usage{
create_arrival_discharge_df(.data)
}
\arguments{
\item{.data}{USe the PTOS_DF}
}
\value{
Normalized dataframe
}
\description{
Create Normalized Arrival and Discharge Data
}
