% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Complications.R
\name{create_complic_df}
\alias{create_complic_df}
\title{Create Normalized Complications Data}
\usage{
create_complic_df(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Create Normalized Complications Data
}
