% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Diagnosis.R
\name{create_diag_df}
\alias{create_diag_df}
\title{Create Normalized Diagnosis Data}
\usage{
create_diag_df(.data)
}
\arguments{
\item{ptos_df}{}
}
\value{

}
\description{
Create Normalized Diagnosis Data
}
