% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create Transactional Flat DF.R
\name{create_flat_trans}
\alias{create_flat_trans}
\title{Create Flat Dataframe of all procedures, complications, and diagnosis by ID.
Uses the trans_full_df as its source}
\usage{
create_flat_trans(.full_trans)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Create Flat Dataframe of all procedures, complications, and diagnosis by ID.
Uses the trans_full_df as its source
}
