% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create Transactional DF.R
\name{create_full_trans}
\alias{create_full_trans}
\title{Combines all the transactional data together}
\usage{
create_full_trans(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Combines all the transactional data together
}
