% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Initial Injury.R
\name{create_injury_df}
\alias{create_injury_df}
\title{Create Normalized Injury Data}
\usage{
create_injury_df(.data)
}
\arguments{
\item{.data}{USe the PTOS_DF}
}
\value{
Normalized dataframe
}
\description{
Create Normalized Injury Data
}
