% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create Patient DF.R
\name{create_patient_df}
\alias{create_patient_df}
\title{Create the Patient Data Frame}
\usage{
create_patient_df(.data, df_name = "patient_df")
}
\arguments{
\item{.data}{Raw PTOS Dataframe}

\item{df_name}{Name of Dataframe}
}
\value{
Returns a data frame for patient information
}
\description{
Create the Patient Data Frame
}
