% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create Patient Time Periods.R
\name{create_patient_timeframe}
\alias{create_patient_timeframe}
\title{Create Patient Time Periods and Intervals}
\usage{
create_patient_timeframe(.patient_df)
}
\arguments{
\item{.patient_df}{Patient Dataframe}
}
\value{
Dataframe of dates, times, and period intervals
}
\description{
Create Patient Time Periods and Intervals
}
