% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Transportation.R
\name{create_transportation_df}
\alias{create_transportation_df}
\title{Create Normalized Transportation Data}
\usage{
create_transportation_df(.data)
}
\arguments{
\item{.data}{USe the PTOS_DF}
}
\value{
Normalized dataframe
}
\description{
Create Normalized Transportation Data
}
