% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAGNOSIS.R
\name{diagnosis_lvl_1}
\alias{diagnosis_lvl_1}
\alias{diagnosis_lvl_2}
\title{Diagnosis Performed}
\usage{
diagnosis_lvl_1(col)

diagnosis_lvl_2(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with 'icd9_'}
}
\value{
It translations of the code into human friendly values.
}
\description{
Diagnosis Performed
}
