% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIS_STATUS.R
\name{dis_status}
\alias{dis_status}
\title{Discharge Status}
\usage{
dis_status(col)
}
\arguments{
\item{col}{A column in PTOS data that equals dis_status}
}
\value{
It translations of the code into human friendly values.
}
\description{
Discharge Status
}
