% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02 - Date of Birth Cleanup.R
\name{dob_cleanup}
\alias{dob_cleanup}
\title{Calculating DOB when DOB is missing using AGE MINUS Admission Date}
\usage{
dob_cleanup(age, dob, adm_dt)
}
\arguments{
\item{age}{Raw AGE of Patient}

\item{dob}{Date of Birth}

\item{adm_dt}{Date of Admission (eda_date_a)}
}
\value{
Returns the Date of Birth of Patient
}
\description{
Calculating DOB when DOB is missing using AGE MINUS Admission Date
}
