% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00 - Drop All NULL Columns.R
\name{drop_blank_columns}
\alias{drop_blank_columns}
\title{Drop All Columns that contain only NA's, <n/a>, <unk>, NULL, etc}
\usage{
drop_blank_columns(.data)
}
\arguments{
\item{.data}{PTOS Data}
}
\value{
Cleaned PTOS Data
}
\description{
Drop All Columns that contain only NA's, <n/a>, <unk>, NULL, etc
}
