% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRUG.R
\name{drug}
\alias{drug}
\title{Record of drugs patient tested positive at referring hospital - Pg. 42}
\usage{
drug(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with DRUG_}
}
\value{
It translations of the code into human friendly values.
}
\description{
Record of drugs patient tested positive at referring hospital - Pg. 42
}
