% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_CODE.R
\name{e_code}
\alias{e_code}
\title{Cause of Injury}
\usage{
e_code(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with E_ACT_1}
}
\value{
It translations of the code into human friendly values.
}
\description{
Cause of Injury
}
