% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECODE_ACT.R
\name{ecode_act}
\alias{ecode_act}
\title{Activity Code for Injury}
\usage{
ecode_act(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with E_ACT_1}
}
\value{
It translations of the code into human friendly values.
}
\description{
Activity Code for Injury
}
