% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EYE_OPNG.R
\name{eye_opng}
\alias{eye_opng}
\title{GCS Assessment of the stimulus required to induce eye opening,
Assessed within 30 minutes or less of ED/hospital arrival  (pg. 67)}
\usage{
eye_opng(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with EYE_OPNG_}
}
\value{
It translations of the code into human friendly values.
}
\description{
GCS Assessment of the stimulus required to induce eye opening,
Assessed within 30 minutes or less of ED/hospital arrival  (pg. 67)
}
