% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM.R
\name{fim}
\alias{fim}
\title{Functional Status at Discharge}
\usage{
fim(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with D_DISABLE_}
}
\value{
It translations of the code into human friendly values.
}
\description{
Functional Status at Discharge
}
