% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.r
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Check if Elements Are Not in a Vector}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector of values to be checked.}

\item{y}{A vector of values to check against.}
}
\value{
A logical vector of the same length as \code{x}, where \code{TRUE} indicates
the corresponding element in \code{x} is not found in \code{y}, and \code{FALSE} indicates
it is found in \code{y}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\verb{\%not_in\%} has been \strong{deprecated} because base R (>= 4.6.0) now provides the
\verb{\%notin\%} operator.

Please use \verb{\%notin\%} from \code{base} instead.
}
\examples{

# Example vectors
x <- c("apple", "banana", "cherry")
y <- c("banana", "grape")

# Check which elements in `x` are not in `y`
x \%not_in\% y

# Example with numeric values
a <- c(1, 2, 3, 4, 5)
b <- c(2, 4, 6)

a \%not_in\% b

}
\author{
Nicolas Foss, Ed.D., MS
}
