% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGT_FALL.R
\name{hgt_fall}
\alias{hgt_fall}
\title{Fall Height}
\usage{
hgt_fall(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with HGT_FALL}
}
\value{
It translations of the code into human friendly values.
}
\description{
Fall Height
}
