% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INJ_TYPE.R
\name{inj_type}
\alias{inj_type}
\title{Type of Injury (pg. 20)}
\usage{
inj_type(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with INJ_TYPE}
}
\value{
It translations of the code into human friendly values.
}
\description{
Type of Injury (pg. 20)
}
