% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02 - Injury Date Cleanup.R
\name{injury_dt_cleanup}
\alias{injury_dt_cleanup}
\title{Calculating Injury Date when Injury Date is missing using AGE PLUS Date of Birth}
\usage{
injury_dt_cleanup(age, dob, injury_dt)
}
\arguments{
\item{age}{Raw AGE of Patient}

\item{dob}{Date of Birth}

\item{injury_dt}{Date of Injury}
}
\value{
Returns a calculated Injury Date based on the Date of Birth and Age of Patient
}
\description{
Calculating Injury Date when Injury Date is missing using AGE PLUS Date of Birth
}
