% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INJURY_LOC.R
\name{injury_loc}
\alias{injury_loc}
\title{Classification of type of injury (pg. 20)}
\usage{
injury_loc(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with e849_x}
}
\value{
It translations of the code into human friendly values.
}
\description{
Classification of type of injury (pg. 20)
}
