% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INJURY.R
\name{injury_lvl_1}
\alias{injury_lvl_1}
\title{Injury ECode Level 1}
\usage{
injury_lvl_1(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with e_code}
}
\value{
It translations of the code into human friendly values.
}
\description{
Injury ECode Level 1
}
