% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIFE_SUP.R
\name{life_sup}
\alias{life_sup}
\title{Highest Level of Provider - Life Support (pg 25)}
\usage{
life_sup(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with LIFE_SUP}
}
\value{
It translations of the code into human friendly values.
}
\description{
Highest Level of Provider - Life Support (pg 25)
}
