% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOT_RESP.R
\name{mot_resp}
\alias{mot_resp}
\title{GCS Assessment of the stimulus necessary to elicit motor response,
Assessed within 30 minutes or less of ED/hospital arrival  (pg. 66)}
\usage{
mot_resp(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with MOT_RESP}
}
\value{
It translations of the code into human friendly values.
}
\description{
GCS Assessment of the stimulus necessary to elicit motor response,
Assessed within 30 minutes or less of ED/hospital arrival  (pg. 66)
}
