% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAY_CAT.R
\name{pay_cat}
\alias{pay_cat}
\title{Pay categories are based on HC4 standard categories (pg. 101)}
\usage{
pay_cat(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with PAY_CAT}
}
\value{
It translations of the code into human friendly values.
}
\description{
Pay categories are based on HC4 standard categories (pg. 101)
}
