% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PGY.R
\name{pgy}
\alias{pgy}
\title{Post Graduate Year for Resident}
\usage{
pgy(col)
}
\arguments{
\item{col}{A column in PTOS data that typically ends in _PGY.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Post Graduate Year for Resident
}
