% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_ED_D.R
\name{post_ed_d}
\alias{post_ed_d}
\title{Final destination after Emergency Department (Pg. 52)}
\usage{
post_ed_d(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with POST_ED_D}
}
\value{
It translations of the code into human friendly values.
}
\description{
Final destination after Emergency Department (Pg. 52)
}
