% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROC_LOC.R
\name{proc_lc}
\alias{proc_lc}
\title{Location of Medical Procedures}
\usage{
proc_lc(col)
}
\arguments{
\item{col}{A column in PTOS data that typically ends with _lc}
}
\value{
It translations of the code into human friendly values.
}
\description{
Location of Medical Procedures
}
\examples{
procedure_location(proc_01_lc)
}
