% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROCEDURE.R
\name{procedure_lvl_1}
\alias{procedure_lvl_1}
\alias{procedure_lvl_2}
\title{Procedures Performed
Record both operative and non-operative procedures performed in your facility}
\usage{
procedure_lvl_1(col)

procedure_lvl_2(col)
}
\arguments{
\item{col}{A column in PTOS data that typically ends in _PR.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Procedures Performed
Record both operative and non-operative procedures performed in your facility
}
