% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROT_DEV.R
\name{prot_dev}
\alias{prot_dev}
\title{Protective devices (safety equipment) in use or worn by the patient at the time of the injury}
\usage{
prot_dev(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with PROT_DEV.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Protective devices (safety equipment) in use or worn by the patient at the time of the injury
}
