% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Column Renaming.R
\name{ptos_rename_columns}
\alias{ptos_rename_columns}
\title{Renaming PTOS Data}
\usage{
ptos_rename_columns(.data)
}
\arguments{
\item{.data}{A dataframe of PTOS data}
}
\value{
It translations of the code into human friendly values.
}
\description{
Renaming PTOS Data
}
