% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACE.R
\name{race}
\alias{race}
\title{Race of Patient using Pg. 17}
\usage{
race(col)
}
\arguments{
\item{col}{Patient Race}
}
\value{
It translations of the code into human friendly values.
}
\description{
Race of Patient using Pg. 17
}
