% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02 - Column Renaming.R
\name{rename_columns}
\alias{rename_columns}
\title{Renaming PTOS Data}
\usage{
rename_columns(.data)
}
\arguments{
\item{.data}{A dataframe of PTOS data}
}
\value{
It translations of the code into human friendly values.
}
\description{
Renaming PTOS Data
}
