% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00 - Run the Full Show.R
\name{run_full_show}
\alias{run_full_show}
\title{Rull the full show for PTOS Data Frame}
\usage{
run_full_show(.data)
}
\arguments{
\item{.data}{PTOS Raw Data}
}
\value{

}
\description{
Rull the full show for PTOS Data Frame
}
