% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SERVICE.R
\name{service}
\alias{service}
\title{Service or Specialty (Not the same as surgical speciality)}
\usage{
service(col)
}
\arguments{
\item{col}{A column typically ends with "_serv"}
}
\value{
It translations of the code into human friendly values.
}
\description{
Service or Specialty (Not the same as surgical speciality)
}
\examples{
service(C01_SERV)
}
