% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEX.R
\name{sex}
\alias{sex}
\title{Classification of Male or Female (pg. 17)}
\usage{
sex(col)
}
\arguments{
\item{col}{Male = 1 and Female = 2}
}
\value{
It translations of the code into human friendly values.
}
\description{
Classification of Male or Female (pg. 17)
}
