% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRANSPORT.R
\name{transport}
\alias{transport}
\title{Transportation (Pg. 24)
Identifies the method used to transport the patient to the trauma center submitting data and the origin of that transport, e.g., accident scene,   rendezvous or transferring hospital.}
\usage{
transport(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with TRANSP_}
}
\value{
It translations of the code into human friendly values.
}
\description{
Transportation (Pg. 24)
Identifies the method used to transport the patient to the trauma center submitting data and the origin of that transport, e.g., accident scene,   rendezvous or transferring hospital.
}
