% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_character_factor.R
\name{validate_character_factor}
\alias{validate_character_factor}
\title{Validate a Character or Factor Input}
\usage{
validate_character_factor(
  input,
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE,
  var_name = NULL
)
}
\arguments{
\item{input}{The data to be validated.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}

\item{var_name}{Optional. A character string giving the desired variable (or
object) name that will appear in console output in place of the how the
object will typically be named in messages via deparse(substitute(input)).}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if an input is of type character or factor. Depending on
the specified type, it will either throw an error, issue a warning, or send a
message. It also checks for NULL and NA values based on the specified
parameters.
}
\examples{
# Synthetic data
data <- data.frame(
  Trauma_Type = c("Blunt", "Penetrating", "Blunt", "Unknown"),
  Patient_Age_Years = c(30, 60, 45, 50),
  RTS = c(7.84, 6.90, 7.00, 6.50),
  ISS = c(10, 25, 15, 20)
)

# Validate the Trauma_Type input
validate_character_factor(
  data$Trauma_Type,
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
