% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_date.R
\name{validate_date}
\alias{validate_date}
\title{Validate Date input}
\usage{
validate_date(
  input,
  na_ok = TRUE,
  null_ok = TRUE,
  type = c("error", "warning", "message")
)
}
\arguments{
\item{input}{The data to be validated.}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if an input contains valid dates and optionally checks
if the values are within a specified range. Depending on the specified type,
it will either throw an error, issue a warning, or send a message. Additional
arguments allow for checking NA values and NULL values.
}
\examples{
# Synthetic data
data <- data.frame(
  Date_input = as.Date(c("2023-01-01", "2023-06-15", "2023-12-31", NA))
)

# Validate the Date_input
validate_date(data$Date_input,
  min = as.Date("2023-01-01"),
  max = as.Date("2023-12-31"),
  na_ok = TRUE,
  null_ok = FALSE,
  type = "warning"
)

}
\author{
Nicolas Foss, Ed.D., MS
}
