% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_date_time.R
\name{validate_date_time}
\alias{validate_date_time}
\title{Validate a Date-Time input}
\usage{
validate_date_time(
  input,
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE
)
}
\arguments{
\item{input}{The data to be validated.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if an input contains valid date-time objects. Depending
on the specified type, it will either throw an error, issue a warning, or
send a message. It also checks for NULL and NA values based on the specified
parameters.
}
\examples{
# Synthetic data
data <- data.frame(
  Event_Time = c(
    "2023-01-01 12:00:00",
    "2023-01-02 13:00:00",
    NA,
    "2023-01-04 15:00:00"
  )
)

# Validate the Event_Time input
validate_date_time(
  data$Event_Time,
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
