% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_error_type.R
\name{validate_error_type}
\alias{validate_error_type}
\title{Validate Error Type}
\usage{
validate_error_type(
  input,
  message,
  type = c("error", "warning", "message"),
  calls = NULL
)
}
\arguments{
\item{input}{The data to be validated.}

\item{message}{The message to be displayed.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{calls}{Optional. The number of callers to go back in the call stack for
error messaging. If NULL, will default to 2.}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function displays an error, warning, or message based on the specified
type.
}
