% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_names.R
\name{validate_names}
\alias{validate_names}
\title{Validate Column Names}
\usage{
validate_names(
  input,
  check_names,
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE
)
}
\arguments{
\item{input}{A data.frame or tibble. validate_names() will run
colnames(input) to get the expected column names.}

\item{check_names}{A vector of column names as strings to check against
input.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if all column names of a data frame or tibble are within
a specified set of valid values. Depending on the specified type, it will
either throw an error, issue a warning, or send a message.
}
\examples{
# Synthetic data
data <- data.frame(
  Trauma_Type = c("Blunt", "Penetrating", "Blunt", "Unknown"),
  Patient_Age_Years = c(30, 60, 45, 50),
  RTS = c(7.84, 6.90, 7.00, 6.50),
  ISS = c(10, 25, 15, 20)
)

# Validate the column names
validate_names(
  input = data,
  check_names = c("Trauma_Type", "Patient_Age_Years", "RTS", "ISS"),
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
