% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_non_empty.R
\name{validate_non_empty}
\alias{validate_non_empty}
\title{Validate Non-Empty Input}
\usage{
validate_non_empty(
  input,
  data_type = c("data.frame", "tibble", "matrix", "list", "array", "atomic vector",
    "numeric", "integer", "logical", "character", "factor", "complex", "raw", "tbl_df",
    "tbl"),
  logic = c("and", "or"),
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE
)
}
\arguments{
\item{input}{The data to be validated.}

\item{data_type}{A vector of data structure or class types to check. Possible
values are "data.frame", "tibble", "matrix", "list", "array", "atomic
vector", "numeric", "integer", "logical", "character", "factor", "complex",
"raw", "tbl_df", "tbl".}

\item{logic}{The logical operator to use when combining checks. Possible
values are "and" or "or".}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if the input is non-empty. Depending on the specified
type, it will either throw an error, issue a warning, or send a message. It
also checks for NULL and NA values based on the specified parameters.
}
\examples{
# Synthetic data
data <- data.frame(
  a = 1:5,
  b = 1:10,
  c = 1:15
)

# Validate the data frame
validate_non_empty(
  data,
  data_type = c("data.frame", "tibble"),
  logic = "or",
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

# Validate an empty vector
validate_non_empty(
  character(0),
  data_type = c("vector"),
  logic = "or",
  type = "error",
  na_ok = TRUE,
  null_ok = TRUE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
