% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_string_length.R
\name{validate_string_length}
\alias{validate_string_length}
\title{Validate String Length}
\usage{
validate_string_length(
  input,
  min_length,
  max_length,
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE,
  var_name = NULL
)
}
\arguments{
\item{input}{The data to be validated.}

\item{min_length}{The minimum length of the strings.}

\item{max_length}{The maximum length of the strings.}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}

\item{var_name}{Optional. A character string giving the desired variable (or
object) name that will appear in console output in place of the how the
object will typically be named in messages via deparse(substitute(input)).}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if the length of strings in an input is within a
specified range. Depending on the specified type, it will either throw an
error, issue a warning, or send a message. It also checks for NULL and NA
values based on the specified parameters.
}
\examples{
# Synthetic data
data <- data.frame(
  Names = c("Alice", "Bob", "Charlie", "David")
)

# Validate the Names input
validate_string_length(
  data$Names,
  min_length = 3,
  max_length = 7,
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
