% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERB_RESP.R
\name{verb_resp}
\alias{verb_resp}
\title{GCS Assessment of the stimulus required for verbal response (or written response if verbal response is impaired by intubation or tracheostomy),
assessed within 30 minutes or less of ED/hospital arrival.  (pg. 67)}
\usage{
verb_resp(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with VER_RESP}
}
\value{
It translations of the code into human friendly values.
}
\description{
GCS Assessment of the stimulus required for verbal response (or written response if verbal response is impaired by intubation or tracheostomy),
assessed within 30 minutes or less of ED/hospital arrival.  (pg. 67)
}
