% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XFER_TIME.R
\name{xfer_time}
\alias{xfer_time}
\title{Elapsed time from date/time of arrival at referring facility to date/time of ED admission (Pg. 54)}
\usage{
xfer_time(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with XFER_TIME}
}
\value{
It translations of the code into human friendly values.
}
\description{
Elapsed time from date/time of arrival at referring facility to date/time of ED admission (Pg. 54)
}
