% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YES_NO.R
\name{yes_no}
\alias{yes_no}
\title{Yes and No Classification}
\usage{
yes_no(col)
}
\arguments{
\item{col}{A column in PTOS data that requires conversion from 1's and 0's to Yes and No.}
}
\value{
It translations of the code into human friendly values.
}
\description{
Yes and No Classification
}
