% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create Patient DF.R
\name{create_patient_df}
\alias{create_patient_df}
\title{Create the Patient Data Frame}
\usage{
create_patient_df(.data, .trans_flat_df, .trans_full_df)
}
\arguments{
\item{.data}{Raw PTOS Dataframe}

\item{.trans_flat_df}{Flat Transactional Dataframe}

\item{.trans_full_df}{Full Transactional Dataframe}
}
\value{
Returns a data frame for patient information
}
\description{
Create the Patient Data Frame
}
