% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03 - Trans Procedure.R
\name{create_proc_df}
\alias{create_proc_df}
\title{Create Normalized Procedure Data}
\usage{
create_proc_df(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
Create Normalized Procedure Data
}
