% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01 - Create PTOS Data.R
\name{create_ptos_data}
\alias{create_ptos_data}
\title{Create Human Friendly PTOS Data}
\usage{
create_ptos_data(.data)
}
\arguments{
\item{.data}{Imported and Raw PTOS Data}
}
\value{
A human friendly data set with normalized ouputs for patient demographics, and time series of medical events
}
\description{
Create Human Friendly PTOS Data
}
