% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOB_Cleanup.R
\name{dob_cleanup}
\alias{dob_cleanup}
\title{Calculating DOB when DOB is missing using AGE MINUS Injury Date}
\usage{
dob_cleanup(age, dob)
}
\arguments{
\item{age}{Raw AGE of Patient}

\item{dob}{Date of Birth}
}
\value{
Returns the Date of Birth of Patient
}
\description{
Calculating DOB when DOB is missing using AGE MINUS Injury Date
}
