% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_it_normal.R
\name{is_it_normal}
\alias{is_it_normal}
\title{Assess Normality of a Continuous Variable}
\usage{
is_it_normal(
  df,
  ...,
  seed = 10232015,
  normality_test = c("shapiro-wilk", "kolmogorov-smirnov", "anderson-darling",
    "lilliefors", "cramer-von mises"),
  include_plots = FALSE,
  plot_theme = traumar::theme_cleaner
)
}
\arguments{
\item{df}{A \code{data.frame} or \code{tibble}. The dataset containing the variable.}

\item{...}{Additional arguments passed to the \code{plot_theme}.}

\item{seed}{A numeric value to set the random seed. Defaults to \code{10232015}.}

\item{normality_test}{A string indicating which test to use. Must be one of:
\code{"shapiro-wilk"}, \code{"kolmogorov-smirnov"}, \code{"anderson-darling"}, \code{"lilliefors"},
or \code{"cramer-von mises"}. Set to \code{NULL} to skip the test.}

\item{include_plots}{Logical. If \code{TRUE}, includes a set of diagnostic plots.}

\item{plot_theme}{A function for setting the \code{ggplot2} theme. Defaults to
\code{traumar::theme_cleaner}.}

\item{x}{A bare (unquoted) column name in \code{df} specifying the continuous
variable to assess.}
}
\value{
A named list with the following elements:
\describe{
\item{descriptive_statistics}{A tibble summarizing the variable's
distribution.}
\item{normality_test}{(Optional) A tidy result of the specified normality
test.}
\item{plots}{(Optional) A \code{patchwork} object containing four
visualizations: Q-Q plot, histogram, density plot, and boxplot with
scatter.}
}
}
\description{
Evaluates whether a numeric variable in a data frame follows a normal
distribution. Provides descriptive statistics, optional visualizations, and
formal normality tests.
}
\author{
Nicolas Foss Ed.D., MS
}
