% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_linear_bins.R
\name{nonlinear_bins}
\alias{nonlinear_bins}
\title{Create Nonlinear Probability of Survival Bins}
\usage{
nonlinear_bins(
  data,
  Ps_col,
  divisor1 = 5,
  divisor2 = 5,
  threshold_1 = 0.9,
  threshold_2 = 0.99
)
}
\arguments{
\item{data}{A `data.frame` or `tibble` containing the probability of survival
data for a set of patients.}

\item{Ps_col}{The column in `data` containing the probability of survival
values for a set of patients.}

\item{divisor1}{A parameter to control the width of the probability of
survival range bins. Affects the creation of step sizes for the beginning
of each bin range. Defaults to `5`.}

\item{divisor2}{A parameter to control the width of the probability of
survival range bins. Affects the creation of step sizes for the beginning
of each bin range. Defaults to `5`.}

\item{threshold_1}{A parameter to decide where data indices will begin to
create step sizes. Defaults to `0.9`.}

\item{threshold_2}{A parameter to decide where data indices will end to
create step sizes. Defaults to `0.99`.}
}
\value{
A list with `intervals` and `bin_stats` objects:
* `intervals`: A vector of start and end-points for the probability of survival bin ranges.
* `bin_stats`: A `tibble` with columns `bin_number`, `bin_start`, `bin_end`, `mean`, `sd`, `count`, and `percent`.
}
\description{
This function generates nonlinear bins for probability of survival data based
on specified thresholds and divisors. It calculates bin statistics, including
mean, standard deviation, count, and percentage for each bin.
}
\examples{
# Generate example data with high negative skewness
set.seed(123)

# Parameters
n_patients <- 10000  # Total number of patients

# Skewed towards higher values
Ps <- plogis(rnorm(n_patients, mean = 2, sd = 1.5))

# Simulate survival outcomes based on Ps
survival_outcomes <- rbinom(n_patients,
                            size = 1,
                            prob = Ps
                            )

# Create data frame
data <- data.frame(Ps = Ps, survival = survival_outcomes) |>
dplyr::mutate(death = dplyr::if_else(survival == 1, 0, 1))

# Apply the nonlinear_bins function
results <- nonlinear_bins(data = example_data,
                          Ps_col = Ps,
                          divisor1 = 5,
                          divisor2 = 5,
                          threshold_1 = 0.9,
                          threshold_2 = 0.99)

# View results
results$intervals
results$bin_stats

}
\author{
Nicolas Foss, Ed.D, MS, original paper and code in MATLAB by Nicholas
  J. Napoli, Ph.D., MS
}
