% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRE_EXIST.R
\name{pre_exist}
\alias{pre_exist}
\title{Pre-existing Medical Conditions}
\usage{
pre_exist(col)
}
\arguments{
\item{col}{A column in PTOS data that typically starts with PEC_}
}
\value{
It translations of the code into human friendly values.
}
\description{
Pre-existing Medical Conditions
}
