% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REF_INTD.R
\name{ref_intd}
\alias{ref_intd}
\title{INTERVENTIONS FOR DIAGNOSIS AND TREATMENTS}
\usage{
ref_intd(col)
}
\arguments{
\item{col}{A column that typically starts with "REF_INTD" or "REF_INTT"}
}
\value{
It translations of the code into human friendly values.
}
\description{
INTERVENTIONS FOR DIAGNOSIS AND TREATMENTS
}
