% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_performance.r
\name{trauma_performance}
\alias{trauma_performance}
\title{Calculate Trauma Hospital Performance Based on Robust and Validated Measures}
\usage{
trauma_performance(
  df,
  Ps_col,
  outcome_col,
  outcome = 1,
  z_method = c("survival", "mortality"),
  diagnostics = FALSE
)
}
\arguments{
\item{df}{A data frame containing patient data.}

\item{Ps_col}{The name of the column containing the probability of survival
(Ps). The values should be numeric and between 0 and 1. Values greater than
1 will be automatically converted to decimal format by dividing by 100.}

\item{outcome_col}{The name of the column containing the binary outcome data.

The column should contain binary values indicating the patient outcome.
Valid values include 1 (dead) and 0 (alive), or TRUE (dead) and FALSE
(alive), or other similar binary representations (e.g., "Yes" for dead and
"No" for alive). The function will check for two unique values in this
column and expects them to represent the outcome in a binary form.}

\item{outcome}{The value representing mortality (default is 1). Can also be set
to 0 or TRUE/FALSE, depending on how the outcome is encoded in
\code{outcome_col}.}

\item{z_method}{A character vector indicating which method to use for
calculating the Z-score. Must be one of "survival" or "mortality". The
default is "survival".}

\item{diagnostics}{A logical flag (default is FALSE). If TRUE, diagnostic
information about the W, M, and Z scores will be printed to the console.}
}
\value{
A tibble containing the calculated scores (W-score, M-score, Z-score)
and associated performance information. If \code{diagnostics} is TRUE,
additional diagnostic information will be printed.
}
\description{
This function calculates trauma hospital performance based on the M, W, and Z
scores, which are derived from survival probability and mortality data, using
established methods (Flora, 1978; Boyd et al., 1987). It
computes the W-score, M-score, and Z-score based on the provided dataset and
calculates performance metrics for trauma programs.
}
\examples{
# Generate example data with high negative skewness
set.seed(123)

# Parameters
n_patients <- 10000  # Total number of patients

# Generate survival probabilities (Ps) using a logistic distribution
set.seed(123)  # For reproducibility
Ps <- plogis(rnorm(n_patients, mean = 2, sd = 1.5))  # Skewed towards higher values

# Simulate survival outcomes based on Ps
survival_outcomes <- rbinom(n_patients, size = 1, prob = Ps)

# Create data frame
data <- data.frame(Ps = Ps, survival = survival_outcomes) |>
dplyr::mutate(death = dplyr::if_else(survival == 1, 0, 1))

# Calculate trauma performance (W, M, Z scores)
trauma_performance(data, Ps_col = Ps, outcome_col = death)

}
\author{
Nicolas Foss, Ed.D., MS
}
