% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trauma_performance.r
\name{trauma_performance}
\alias{trauma_performance}
\title{Calculate Trauma Hospital Performance Based on Robust and Validated Measures}
\usage{
trauma_performance(
  df,
  Ps_col,
  outcome_col,
  outcome = 1,
  z_method = c("survival", "mortality"),
  diagnostics = FALSE
)
}
\arguments{
\item{df}{A data frame containing patient data.}

\item{Ps_col}{The name of the column containing the probability of survival (Ps). The values
should be numeric and between 0 and 100 (percent). Values greater than 1 will be automatically
converted to decimal format by dividing by 100.}

\item{outcome_col}{The name of the column containing the binary outcome data.

The column should contain binary values indicating the patient outcome.
Valid values include 1 (dead) and 0 (alive), or TRUE (dead) and FALSE (alive),
or other similar binary representations (e.g., "Yes" for dead and "No" for alive).
The function will check for two unique values in this column and expects them to represent
the outcome in a binary form.}

\item{outcome}{The value representing death (default is 1). Can also be set to 0 or TRUE/FALSE,
depending on how the outcome is encoded in `outcome_col`.}

\item{z_method}{A character vector indicating which method to use for calculating the Z-score.
Must be one of "survival" or "mortality". The default is "survival".}

\item{diagnostics}{A logical flag (default is FALSE). If TRUE, diagnostic information about
the W, M, and Z scores will be printed to the console.}
}
\value{
A tibble containing the calculated scores (W-score, M-score, Z-score) and associated
performance information. If `diagnostics` is TRUE, additional diagnostic information will be printed.
}
\description{
This function calculates trauma hospital performance based on the M, W, and Z scores,
which are derived from survival probability and mortality data, using established methods
(Flora, 1978; Boyd et al., 1987; Mock et al., 2004). It computes the W-score, M-score,
and Z-score based on the provided dataset and calculates performance metrics for trauma programs.
}
\details{
The function computes three scores:
\itemize{
  \item W-score: A measure of the hospital's performance compared to expected survival.
  \item M-score: A measure of the patient case mix similarity to the Major Trauma Outcomes Study (MTOS).
  \item Z-score: A measure of the difference between observed and predicted outcomes, based on the selected method.
}

The function checks the validity of the data columns and alerts the user if any issues are detected.
}
\author{
Nicolas Foss, Ed.D., MS
}
