% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traumar-package.R
\docType{package}
\name{traumar-package}
\alias{traumar}
\alias{traumar-package}
\title{traumar: Calculate Metrics for Trauma System Performance}
\description{
Hospitals, hospital systems, and even trauma systems that provide care to injured patients may not be aware of robust metrics that can help gauge the efficacy of their programs in saving the lives of injured patients. 'traumar' provides robust functions driven by the academic literature to automate the calculation of relevant metrics to individuals desiring to measure the performance of their trauma center or even a trauma system. 'traumar' also provides some helper functions for the data analysis journey.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bemts-hhs.github.io/traumar/}
  \item \url{https://github.com/bemts-hhs/traumar}
  \item Report bugs at \url{https://github.com/bemts-hhs/traumar/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Foss \email{nicolas.foss@hhs.iowa.gov}

Other contributors:
\itemize{
  \item Iowa Department of Health and Human Services [copyright holder]
}

}
\keyword{internal}
