% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_pull.R
\name{validate_data_pull}
\alias{validate_data_pull}
\title{Validate Data Extraction}
\usage{
validate_data_pull(data, col, var_name = NULL, calls = NULL)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{col}{The column to be extracted.}

\item{var_name}{Optional. The name of the variable for error messaging.}

\item{calls}{Optional. The number of callers to go back in the call stack for
error messaging. If NULL, will default to 5.}
}
\value{
The extracted column as a vector.
}
\description{
This function extracts a column from a data frame or tibble and returns it as
a vector. If the column does not exist or an error occurs, it returns a clean
error message using the cli package.
}
\examples{
# Synthetic data
data <- data.frame(
  Trauma_Type = c("Blunt", "Penetrating", "Blunt", "Unknown"),
  Patient_Age_Years = c(30, 60, 45, 50),
  RTS = c(7.84, 6.90, 7.00, 6.50),
  ISS = c(10, 25, 15, 20)
)

# Validate and extract the Trauma_Type column
validate_data_pull(
  data,
  Trauma_Type,
  var_name = "Trauma_Type",
  calls = 4
)

}
\author{
Nicolas Foss, Ed.D., MS
}
