% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_structure.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate Data Structure}
\usage{
validate_data_structure(
  input,
  structure_type = c("data.frame", "matrix", "list", "array", "atomic_vector", "tbl_df",
    "tbl"),
  logic = c("and", "or"),
  type = c("error", "warning", "message"),
  na_ok = TRUE,
  null_ok = TRUE,
  var_name = NULL
)
}
\arguments{
\item{input}{The data to be validated.}

\item{structure_type}{A vector of data structure types to check. Possible
values are "data.frame", "matrix", "list", "array", "atomic_vector",
"tbl_df", "tbl".}

\item{logic}{The logical operator to use when combining checks. Possible
values are "and" or "or".}

\item{type}{A character string specifying the type of message to be displayed
if the input is not numeric or if the values are out of range. Must be one
of "error", "warning", or "message".}

\item{na_ok}{Logical. If TRUE, NA values are allowed. Default is TRUE.}

\item{null_ok}{Logical. If TRUE, NULL values are allowed. Default is TRUE.}

\item{var_name}{Optional. A character string giving the desired variable (or
object) name that will appear in console output in place of the how the
object will typically be named in messages via deparse(substitute(input)).}
}
\value{
NULL. The function is used for its side effects.
}
\description{
This function checks if an input is of the specified data structure type(s).
Depending on the specified type, it will either throw an error, issue a
warning, or send a message. It also checks for NULL and NA values based on
the specified parameters.
}
\examples{
# Synthetic data
data <- data.frame(
  a = 1:5,
  b = 1:10,
  c = 1:15
)

# Validate the data structure
validate_data_structure(
  data,
  structure_type = c("data.frame", "tbl", "tbl_df"),
  logic = "or",
  type = "warning",
  na_ok = FALSE,
  null_ok = FALSE
)

}
\author{
Nicolas Foss, Ed.D., MS
}
